+++ tomate/filegroup/filegroup_netcdf.py
@@ -8,7 +8,7 @@
 
 import logging
 import os
-from typing import Any, List
+from typing import Any, Dict, List
 
 try:
     import netCDF4 as nc
@@ -20,7 +20,7 @@ else:
 import numpy as np
 
 from tomate.accessor import Accessor
-from tomate.custom_types import File
+from tomate.custom_types import File, KeyLike
 from tomate.keys.keyring import Keyring
 from tomate.filegroup.filegroup_load import FilegroupLoad
 from tomate.filegroup.command import separate_variables, Command
@@ -105,14 +105,48 @@ class FilegroupNetCDF(FilegroupLoad):
         order = list(file[ncname].dimensions)
         return order
 
-    def write(self, filename: str, wd: str, keyring: Keyring):
-        """Write data to disk."""
+    def write(self, filename: str, wd: str = None,
+              file_kw: Dict = None, var_kw: Dict[str, Dict] = None,
+              keyring: Keyring = None, **keys: KeyLike):
+        """Write data to disk.
+
+        The filegroups should have their variable scanning
+        coordinate setup correctly. See :func:`FilegroupLoad.write
+        <tomate.filegroup.filegroup_load.FilegroupLoad.write>`.
+
+        Variable specific arguments are passed to `netCDF.Dataset.createVariable
+        <https://unidata.github.io/netcdf4-python/netCDF4/index.html
+        #netCDF4.Dataset.createVariable>`__. If the 'datatype' argument is
+        not specified, the 'datatype' attribute is looked in the VI, and if
+        not defined, it is guessed from the numpy array dtype.
+
+        If the 'fill_value' attribute is not specifed, the '_FillValue'
+        attribute is looked in the VI, and if not defined
+        `netCDF4.default_fillvals(datatype)` is used. It seems preferable
+        to specify a fill_value rather than None.
+
+        All attributes from the VariablesInfo are put in the file if their
+        name do not start with an '_'.
+
+        :param wd: Directory to place the file. If None, the
+            filegroup root is used instead.
+        :param file_kw: Keywords argument to pass to `open_file`.
+        :param var_kw: Variables specific arguments.
+        """
         if wd is None:
             wd = self.root
-
         filename = os.path.join(wd, filename)
 
-        with self.open_file(filename, mode='w', log_lvl='INFO') as file:
+        keyring = Keyring.get_default(keyring=keyring, **keys)
+
+        if file_kw is None:
+            file_kw = {}
+        if var_kw is None:
+            var_kw = {}
+
+        file_kw.setdefault('mode', 'w')
+        file_kw.setdefault('log_lvl', 'INFO')
+        with self.open_file(filename, **file_kw) as file:
             for name, coord in self.db.loaded.coords.items():
                 key = keyring[name].copy()
                 key.set_shape_coord(coord)
@@ -126,31 +160,47 @@ class FilegroupNetCDF(FilegroupLoad):
                     file[name].setncattr('fullname', coord.fullname)
                     file[name].setncattr('units', coord.units)
 
-            for info in self.db.vi.infos:
-                if not info.startswith('_'):
-                    file.setncattr(info, self.db.vi.get_info(info))
-
             for var in keyring['var']:
                 cs = self.cs['var']
                 name = cs.in_idx[cs.idx(var)]
-                t = self.vi.get_attr_safe('nctype', var, 'f')
+
                 dimensions = keyring.get_non_zeros()
                 dimensions.remove('var')
 
-                fillvalue = self.db.vi.get_attr_safe('_FillValue', var)
-                if fillvalue is None:
-                    dtype = self.db.data.dtype
-                    tp = '{}{}'.format(dtype.kind, dtype.itemsize)
-                    fillvalue = nc.default_fillvals.get(tp, None)
+                kwargs = var_kw.get(var, {})
 
-                file.createVariable(name, t, dimensions, fill_value=fillvalue)
+                datatype = kwargs.pop('datatype', None)
+                if datatype is None:
+                    datatype = self.vi.get_attr_safe('datatype', var, None)
+                    if datatype is None:
+                        dtype = self.db.data.dtype
+                        datatype = '{}{}'.format(dtype.kind, dtype.itemsize)
+
+                if 'fill_value' not in kwargs:
+                    if '_FillValue' in self.vi[var]:
+                        kwargs['fill_value'] = self.vi[var]._FillValue
+                    else:
+                        kwargs['fill_value'] = nc.default_fillvals.get(datatype, None)
+
+                file.createVariable(name, datatype, dimensions, **kwargs)
                 file[name][:] = self.db.view(keyring=keyring, var=var)
 
-                if var in self.db.vi.variables:
-                    attrs = self.db.vi[var]
-                    for attr in attrs:
-                        if not attr.startswith('_'):
-                            file[name].setncattr(attr, self.db.vi.get_attr(attr, var))
+            self._add_vi_to_file(file, keyring)
+
+    def _add_vi_to_file(self, file, keyring):
+        """Add metadata from VI to file."""
+        for info in self.db.vi.infos:
+            if not info.startswith('_'):
+                file.setncattr(info, self.db.vi.get_info(info))
+
+        for var in keyring['var']:
+            cs = self.cs['var']
+            name = cs.in_idx[cs.idx(var)]
+            if var in self.db.vi.variables:
+                attrs = self.db.vi[var]
+                for attr in attrs:
+                    if not attr.startswith('_'):
+                        file[name].setncattr(attr, self.db.vi.get_attr(attr, var))
 
     def write_variable(self, file: 'nc.Dataset', cmd: Command,
                        var: str, inf_name: str):